runtime dependency bootstrap script
bootstrap_runtime will use virtualenv and pip to install all required python
libraries (see runtime.pip.req) into the runtime directory. The python binary in
runtime can be used to run the server as needed. This is not required but can
simplify new deployements.
Note: This does not check, install, or run redis-server but will install python
bindings for the client.