Fix for inactive hacker-delete button
Refreshing page used to cause the hacker delete button to become
inactive. This was because each hacker's view was not being loaded
along with the page, which meant that none of the view's javascript
hooks were in place.

The attempted optimisation was to avoid a backbone.js ajax call after
page load. But said ajax call is now back in place because I couldn't
figure out how else to do this.

Fixes #14